Fix appliance catalog search bugs
The appliance catalog search had a couple different bugs at once. First, the and/or was not being applied correctly nor having the expected results. AND/OR now applied to the keyword search and if AND, all keywords must match; if OR any keyword must match. Second, multiple searches in a row were sometimes being applied in series, i.e. refining the search results instead of searching the whole catalog again. Both of these are now fixed.