Update mailman sync to look at only new users
The new sync behavior will be to sync only members who have recently
signed up with Chameleon. This updates the sync endpoint to not be
list-specific, and also to only pull users who have joined since the
beginning of the day. As long as this runs once a day, we shouldn't miss
anybody.