Increase window by 24 hours maximum
Before, the sync would only capture users who signed up between 00:00
and when the script ran on that day. It needs to look further back. We
add an extra 24 hours of buffer. This list of new members can return the
same member on two subsequent days; that is fine and should not really
affect anything downstream.