Update Django packages
This updates all of our Django packages, e.g. django-cms, to the latest
version that supports Django 1.8.

The big changes is that DjangoCMS was updated from 3.1 to 3.5, which
required several changes to how apphooks are registered.

djangocms_blog was moved out to reference an externally maintained fork
of that package.