Remove cmsplugin_filer dependency
This package has been deprecated for a while and has a loose dependency
on filer, which dropped Python2 support recently. Unless we fork this
package, we can't keep using it. It looks like it isn't really used
anymore in later versions of djangocms and filer support has been
integrated as the default handler for file inputs.