Write changes back to origin Keystone server (#92)
This significantly changes how the sync logic works within Portal when a
user is logging. Currently we sync different properties at different
times:

1. The user's enabled/disabled status is synced upon every login to
  Portal (regardless of how initiated)
2. The user's projects are lazily created/enabled/disabled upon every
   login _initiated via Horizon WebSSO_ (an important distinction)
3. The user's memberships in projects are synced upon every login
   initated _via WebSSO_.
4. A PI's membership in a project is synced upon every login of every
   user associated with that project initiated _via WebSSO_.
5. Project nicknames are synced _when they are updated_ in the Portal
   UI, and are not synced during login.

All of those write back to the TACC Keystone "primary" server, and
changes are replicated automatically to UC. In order to turn off
replication, we need to ensure each site is updated separately.

This commit thus adds the following logic:

- When a user is logging in via WebSSO, detect for which region they are
  logging in. This will be passed from Horizon.
- Look up the Keystone URL for that region (configured in Portal) and
  perform any operations directly against that Keystone service.
- Operations that take place "out of band" of a login (e.g., setting a
  project nickname) will update all configured Keystones sequentially.
- When generating unscoped tokens, which require access to the user's
  plaintext password (and therefore only at login time), we generate
  unscoped tokens for each region and store a token for each. This is
  because user IDs may be different on each site, and unscoped tokens
  are bound to user IDs, it turns out.

This should ensure that users are able to access the resources on the
site they are trying to log in to (TACC or UC).

NOTE: currently all of the code assumes that the user's domain in
Keystone will match the service account's domain. This will not be
correct when users are federated (they will be in their own domain), but
all of this code only applies to legacy users anyways, so it's OK.