Don't fetch users from Keycloak when showing projects
This updates the project accessor to not require calling out to Keycloak
to fetch the user's metadata. We have PI status in the DB and the rest
of the Keycloak fields we can do without. We just will not show things
like affiliation on the allocation admin view. The user's email should
give some context as to the affiliation.