Gracefully handle missing Keycloak users
During the migratio period, it is possible for users to not exist in
Keycloak. This is especially true as users will not be added to Keycloak
until they are part of an active allocation. When all users are logging
in via Keycloak this will no longer be the case, and we can assume all
users are existing.