Fixup Angular views in CSP context
When CSP is enabled, Angular needs a special CSS sheet in order to
perform its show/hide logic. We are now using CSP, hence we need this.

I couldn't figure out how the main.css file was being generated, so I
threw it out in favor of the appliance_catalog.css, which was being used
instead of main.css in a few places (implying some confusion/difficulty
getting main.css to work). We now use the asset pipeline to compress and
assemble that file.

.. NOTE::

   The ng-csp directive just avoids Angular trying to detect whether
   CSP is enabled itself (which triggers an error in the console.)