Add toggle for opting in globally to federation
Because it is possible to set cookies on all subdomains by using a
leading '.' on the cookie domain, we can effectively opt a user in to
all web properties via a single endpoint.

This also supports an opt-out flow.

To opt in, visit: /new-login-experience
To opt out, visit: /new-login-experience?opt-out