Combine Keycloak memberships with TAS
This is necessary due to the asynchronous nature of the Keycloak/TAS
sync; it only runs every so often due to latency constraints, yet users
may still be trying to access the system. When we are writing everything
directly to Keycloak this will not be an issue, but for now, we can get
out of sync. Always pulling from both sources is the easiest fix; we can
simply remove this block of code later.