Fix new project create flow
When creating a new project, after the TAS request succeeds, the user is
taken to a confirmation page showing them their project. This however
failed b/c when we try to load this page, we end up trying to re-request
the project and in the process attempt to look up the project by its
charge code. When we do this lookup we search the list of all user
projects, but there's a catch: we can only get a list of the user's
active projects! New projects will obviously not be here.

The "fix" is to know that there's really only 1 user of a
not-yet-approved project, and it is the PI. Just shim them in to the
project list. Hackhackhack.