Add migration page for user/project data
This adds a new static page at /user/migrate, which prompts the user (if
they are federated and have a TAS account linked) to migrate their old
data, including keypairs and disk images.

Most of the interaction takes place in the client browser. This also
comes with a new client-side "framework" using Vue. The old
bower/Grunt/angular based stuff is really old, and it's unclear how to
build it in the future. The Vue approach is a bit simpler; you just drop
*.vue files into a vue/ directory in the app for your components, and
*.js files for the entry points.

The entry points need to be a bit special, see the migrate_account.js
for an example. Mostly, you need to set up XSRF stuff for Django to
allow the client to send XHR.