Implement custom logout + redirect registration
All new registrations will now act as logins... there is no difference
from Portal's PoV if the user is new or returning. IdP is also handling
asssenting to ToS.

This also adds a custom logout method, which will perform a "deep"
logout if the user is federated, also logging them out of the IdP. If
they are not federated, it works as before.