Build client static assets separately
Because the client asset build's Docker build context was the entire
project directory, it would often re-do a lot of work. This moves the
client build out so it only builds off of the files it actually needs
(anything in vue/ directories, plus some top-level config files).

This also removes some of the leftover cruft around the Docker build
that we don't need anymore.

The static assets are now expected to be at /static, and media at
/media. The deploy workflow should be to run collectstatic immediately
upon rollout.