Update progress bar logic
The migration page progress bar is a bit special... this makes it
actually make some sense. It now just returns the current percentage,
and we interpolate based on which region we're working on.

If there are 3 regions/sites, then the first stage will take from
1%-33%, the second from 34%-66%, the last from 67%-100%. The stages
might not take the same amount of time, as it depends on how many images
there are to migrate, which we don't know until we process each
individual region.