lazy add user when log in; use user portal id
1) Users are not added to Keycloak if the user is not part of any active
allocations in TAS, though we have allocations approved for them in portal
DB. Therefore, we add TAS-authenticated users to Keycload when they log in.

2) Projects and allocations are created directly to portal DB. When legacy
users create new allocations, the user id is their TAS id. When using TAS
id as the requestor id and saving the new allocation to the portal DB, it
will cause the foreign key conflict. Therefore, we get the portal user id
as the requestor id.