Change allocation job frequency to 5min
This is not an expensive job and running it faster should limit support
questions around how long it takes for an allocation to be activated.

Also fixed the calculation of the time in minutes. Using straight
division could be a ticking time bomb if the result is not a round
number.