Fix issues requesting PI eligibility
- If the user did not have citizenship defined, the form would fail b/c
  this was a required field but the user could not edit

To fix this, this overhauls the edit form. We now handle empty
citizenship and ask the user here. There were issues mapping Keycloak
countries to the TAS list; drop the TAS list and settle on ISO-3166,
which is used in the Keycloak form. Also stop asking for institutions
and departments, which could not match anymore.