Match user on username
There are some users in the Portal DB who have been previously imported
into Keycloak via TAS, yet have their email updated in Portal somehow.
Possibly they updated it themselves, or else their primary email has
changed. The username is not mutable, so we can use it as the key
when trying to match to an existing user (pre-federation).