Remove TAS for all project/user calls
We were still pushing updates to TAS optimistically. This removes
all of that code. We no longer use TAS to backfill the list of user
projects. If projects are somehow not in Keycloak, this can mean
that some user projects will be removed! We should do a sanity check
on the TAS projects by running the sync to ensure this will be OK.