Fix GA tracking beacons
Google Analytics uses not only script includes, but also both image
and POST-based beacones. `default-src` acts as fallback for all of the
CSP directives and is the "easiest" (and most coarse-grained) way
to go. It may be possible to go with a stricter policy later, like:

  script-src 'self' https://www.google-analytics.com;
  img-src https://www.google-analytics.com www.google-analytics.com https://stats.g.doubleclick.net;
  connect-src https://www.google-analytics.com www.google-analytics.com https://stats.g.doubleclick.net

default-src and script-src are both needed here; default-src only
applies for any CSP directive that is undefined, and we're defining
script-src already.