Remove nonce from GA script
Including the csp_nonce tag somehow triggers the nonce to be included
in the CSP headers; if the tag is not used, the header value is not sent.
Sending a nonce in the CSP header actually _disables_ unsafe-inline,
which breaks all kinds of stuff for us, notably django-cms, which doesn't
have proper support for CSP at all (the toolbar is non-operational.)

There are other cases where things break, like when scripts normally
bundled together are unbundled and for some reason no nonce is included
in them ;_;

What a rabbit hole.