enable lint and autoformat pyproject.toml
pyproject.toml is defined in pep621 for project metadata.

This will provide a single location to define config for
linters, autoformatters, and build tools.

The goal is to ensure that local editing and the repo hooks
reflect the same state, and allow us to converge on some style
guidelines.

This also includes a github action to run these checks.