Enable resetting legacy TAS passwords
Currently when a user wants to reset their TAS password, they go to
the TACC user portal to do so. However, for reasons unknown, TAS will
send them an email with a link to a page on the Chameleon user portal.
Probably this was a UX optimization that is no longer relevant.

For now, we simply direct back to the TACC user portal for this.