change middleware syntax for 2.0
Django 2.0 deprecates the 1.9 middleware syntax. All builtin middleware
is compatible with both the old and new styles.

A compat mixin is provided to upgrade old custom middleware, and this
was added to usernews.

The sessionauth middleware is now implicit, and cannot be disabled.
As part of that changeset, it now returns bools, rather than methods.