Add notion of Chameleon-supported artifacts
An artifact is Chameleon supported if the Chameleon team is reponsible
for maintenance. This is indicated via a special label named "chameleon".
If that label is set, an additional icon is included in the tag labels
UI section. This label can only be set by Portal staff members and
normal users do not even see it.