Fixup RT ticket views
This is a general refactor of the RT app, but just the view layer.
This moves most of the ticket handling logic to a core function, which
is called for both the logged in and guest case. Also fixes a bug where
mimetype was set to a tuple and not a string, which apparently caused
RT to throw its hands in the air and not process the file properly.

Added some better/more consistent error messaging, and ensured that
Python warnings were making it to the logging infra--the RT client uses
this mechanism to emit errors, so it's important we are seeing them.