quick-fix Allow requests from prod's host IP
Requests are coming from the host namespace on chameleon06
The "correct" way to handle this is to fix whatever is sending the
request to use the app docker network

A second issue is our usage of os.envion.get, since it has issues
with non-string data types.

We should use a library like `environs` to handle list types.