Add Openstack information in RT ticket description (#211)
* Add Openstack information in RT ticket description

Project, lease, and instance information is added to the end of
a ticket submitted by a logged in user. If a user is anonymous,
no information is displayed.

The version of python-blazarclient is switched to use Chameleon's
version of this library, so that `list_allocations` can be used.

Also, python-ironicclient is added as a new dependency to get the
provision state and instance uuid of a reserved node. This should
make it easier to find the issue when a node in a user's lease
won't launch.