Warn users when allocation is expiring soon (#217)
Implements a new Celery task in the allocations app which warns PIs when their project's allocation is within 1 month of expiring. The email will only be sent once, meaning that the vast majority of users going forward will receive the email exactly one month out from their allocation's expiration date.

Regardless of when the mail was sent, it will calculate the number of days until the allocation's expiration, and notify the PI of that specific number.

We persist this information in the allocations_allocation table with the new field 'expiration_warning_issued', which is a timestamp (default NULL). PIs will only be sent an email if this value is NULL. This means that the only anticipated circumstance under which someone will receive two warning emails is if there is an error writing the timestamp of their previous email to the database

The expiration warning task only occurs once per day. I chose the hour of 7 AM UTC, since I believe it will be the same day for most of our users (between the US west coast and Japan) and it's during the work day for most of the eastern hemisphere.