Allow deleting entire Trovi artifact
Adds a button to delete all artifact versions, and then the artifact itself. The button is accessible via the edit page for a given artifact. Currently, it will launch a JavaScript prompt that will ask for the name of the artifact to confirm its deletion. It will wait until the correct name is input, or the prompt is cancelled.

This achieves the desired function of the confirmation prompt, but I think it would be more ideal to mimic the behavior of other websites (such as deleting a repo on GitHub) where a prompt pops up, and the confirm button is greyed out until the text in the box matches the desired input. I think this is more familiar to users, but unfortunately seems like a challenging front end task, which I am not very familiar with.

Some additional refactoring to the version deletion code has been done on the backend as well, which should make error reporting for delete_version and delete_all more user-friendly.