display institution and country of the pi
we'd like to grant pi eligibility to international researchers,
but have more restrict rules to approve their allocation requests.
in order to easily tell if the project belongs to an international
pi, the allocation admin page will display the institution and country
of the pi. The information is collected from keycloak.