Add filters for GPU and RDMA features
Expose existing GPU features, namely count, model, and vendor

Add RDMA features based on presence of InfiniBand interface.
Presence of Infiniband is based on at least one network adapter having the property `interface: InfiniBand`. Unfortunately this is a reserved word, so we have to use {interface: iface} to keep JS happy.

GPUDirect is Mellanox/Nvidia's term for RMDA to the GPU memory. This is supported on most Nvidia graphics cards, but I've added a whitelist here for safety. Currently the filter just checks for the presence of both infiniband, and a supported GPU. Technically the RDMA connection can be via a Mellanox Infiniband adapter, or via an Ethernet connection supporting L2 or L3 lossless transport (RoCE / iWARP). Since doing this over Ethernet requires a specific switch setup, so this can't be determined out of band.

NVMeOF is Mellanox/Nvidia's term for RDMA to an NVMe storage device (aka attached directly to the PCIe bus). Again, this is just checking for the presence of both Infiniband and a NVMe drive, but technically an in-band command should be run to verify functionality via loopback, or between two machines.