Fix bug with detail list rendering
This fixes a very odd bug where when you apply different filters, the
hardware list will not appear updated in some cases, will show old data,
or data you do not expect. The root cause was that the :key field in
the list that Vue uses to understand if the item should be re-rendered
only depended on the list index and the length of the list. Not sure what
I was thinking but this is not right at all. Using a combination of the
index and the uid of the node will ensure that whenever the hardware's
position in the list changes, or it's added or removed from the list,
this view will always get re-rendered properly.