Add a series of nested questions to the ticket create page
With the goal of not annoying users or making it harder to submit
tickets (since the help desk is a frequently complimented feature),
this adds a series of questions at the top of the ticket create page
that can help a user solve their own issue.

Ideally, this text will help us by reducing ticket load, and help users
by quickly solving their problems. It also provides us a place to leave
feedback for what information should be included for certain types of
questions.