Add our domains to image csp
Our images are uploaded either while on chameleoncloud.org
or www.chameleoncloud.org. Since the absolute URL is used in the CMS,
the images won't load when using the site via the other domain, since
the CSP only contained 'self' (along with some google domains). This
fix adds both domains, meaning that it should always be able to load the
images.