Reverse decorator ordering
Decorators are often described as working "inside-out", meaning
the one closest to the function def is run first. This is true
in terms of the fact that if you have:

  @f
  @g
  def h()

It logically is composed as f(g(h())), e.g., g wraps h and thus
is invoked "first" to get the response value.

However, ultimately what's happening is function references are
passed in to the decorator wrappers and they decide when the function
they wrapped is invoked. We do a lot of preprocessing before invoking
the function; that pre-processing will always happen in-order, so f's
code runs before g's. For this reason, the earliest checks need to
come first.