Improve Trovi error message reporting
Any unhandled Trovi exceptions in Portal would previously give an opaque error message to the user, which would not inform them of what actually went wrong. Now, any unhandled Trovi errors will have the API error message forwarded to the user.

`TroviException` is been modified to hold on to more error context.

The `handle_trovi_errors` decorator is implemented to wrap views with a `TroviException` handler, so that any `TroviExceptions` which would otherwise be unhandled are caught, and the API error message is reported to the user. This does not impede any explicit error handling within view functions.