Fix some direct error handling
There was some error reporting inside a generator expression that was never actually executed, so errors were never displayed.

The `PATCH` request generated by the sharing form also displayed an opaque error message, so handling has been referred to the decorator. This is the easiest way to handle patch errors, because it will explicitly display the fields which are incorrect.