Add project tags (#334)
This change adds project tags for portal projects, as described in the design doc. Tags replace the old project `Type` model. 

This change does not remove `Type`s, but does remove all references to them. The `Type`s need to stay for the migration to re-label old projects. They will need to be removed in a future change.

This change _does_ remove the field of science from projects, and all references to it. 

Some unused TAS code was removed, in order to make the change easier. 

Some TAS code has been renamed to be more generic, in order to make it more obvious what its function is. A lot of the TAS code was not able to be removed without significantly refactoring the entire project/allocation system. That should happen in a future change. 

The `Tag` model is basically the same as `Type`, but with an added `description` field. This is so we can use the shorthand where desired, or optionally include the long description to provide more context to users.