Add task to automatically tag  (#338)
This patch adds a task to automatically tag any projects tagged as "Other." It will tag the project, mark it as automatically_tagged, and then send an email to the project's PI to inform them that their project has been tagged, and tell them where they can go to change it. Each PI will only receive one email.

The tagging is done with nltk. No machine learning is used, since obtaining training data would be more difficult than just manually tagging the project. The task uses nltk's word normalizing utilities to reduce keywords to their root, and then count matches against a list of keyword roots for each tag. The tag with the most matches is chosen, and the project is marked as automatically tagged.

Against actual project descriptions, this method is quite accurate, but not perfect. To counteract this, we send the emails to the PIs, and we are also waiting until the 16th to run this task. The automatically_tagged attribute is added in case we ever want to filter out any projects that were not tagged by hand, in order to get more accurate metrics or just tag them ourselves.

Other changes:

* Add short circuits for tag migration
    * This is only for empty databases

* Take dependency on `nltk`

* Upgrade to poetry 1.2