Fix DOI display in Artifact detail page (#341)
The artifact detail page was not properly displaying DOIs. This patch fixes that.

It also alters the previous behavior. The detail page used to show the DOI for whichever version was the most recent, regardless of which version was selected. It now shows the DOI for the selected version only.

The DOI check for making an artifact private on the share page was also removed. Artifacts with DOIs are allowed to be private, but versions with DOIs are still shown.