Fix tag widget in artifact edit form
There was previously a bug with the tag widget, which would cause it to initially be empty. Because of this, saving the artifact without any changes would cause the tags to be removed from the artifact. Now, the widget will be initialized with whatever tags the artifact had previously.

The widget has also been changed to a multiple-choice checkbox. The multi-select menu thing seems more difficult to use, as it required users to CTRL+Click on multiple options in a tiny scrolling field. It was also not possible to view all of the tags at once in this view.