Fix Artifact edit form fields
There were some bugs with how the form generated POST request because of some missing or typo'd prefixes. These issues are rectified by this patch.

This patch also fixes an issue with the form handler which generated a PATCH request to Trovi very naively. The handler now checks for empty forms, and attempts to completely remove them rather than just set them to a blank value, which Trovi rejects.