Add featured artifacts to top of Trovi homepage (#399)
The featured artifacts are stored in a special table in the datbase, and boosted to the top of the Trovi homepage. 

New featured artifacts are added by using the Featured Artifacts admin page and pasting a UUID in the "New Featured Artifact" form. If the UUID does not map to a real Trovi artifact, the UUID will be ignored when the Trovi homepage is rendered.