Add admin API for project allocation details
We'd like to use this API in order to check at a Chameleon site if a
project has an active allocation. That allows us to clean up resources
for projects that have expired.

This new API endpoint is similar to one we have to "project extras",
and implements very basic token auth. This endpoint only exposes very
basic information as well, some info about the project (in case a site
wants to contact a PI), and information about the allocations. This
inclues the last date the project had/has an allocation for, if
they currently have an active allocation, and if they have a pending
allocation. This way, we can adjust our implementation as we see fit
around these parameters.