Fix daypass view decorators
The decorator order was wrong for the /share view which allows users to
edit daypass settings. Before, the trovi token in the session was not
set before fetching the artifact. This never came up before the artifact
detail page handled this properly, which was the precursor to /share
page. However, users are not being directed here from the external
dashboard, and so the bug came up.