Add RawPublication model
This includes more info in the source, which can be normalized in the
actually Publication model. We had to introduce a new model for this,
and migrate the PublicationSources to it because the previous
constraints on the old model don't play well with the new data. With
this new model, we get rid of the confusing "duplicate" status.